Move calculate_episode_data_index
