+++ tests/test_sampler.py
@@ -15,9 +15,9 @@
 # limitations under the License.
 from datasets import Dataset
 
+from lerobot.common.datasets.push_dataset_to_hub.utils import calculate_episode_data_index
 from lerobot.common.datasets.sampler import EpisodeAwareSampler
 from lerobot.common.datasets.utils import (
-    calculate_episode_data_index,
     hf_transform_to_torch,
 )
