+++ tests/test_utils.py
@@ -7,8 +7,8 @@ import pytest
 import torch
 from datasets import Dataset
 
+from lerobot.common.datasets.push_dataset_to_hub.utils import calculate_episode_data_index
 from lerobot.common.datasets.utils import (
-    calculate_episode_data_index,
     hf_transform_to_torch,
 )
 from lerobot.common.utils.utils import (
